Simple syntect-based autoindent
This adds support for basic autoindent, for languages supported by
syntect, through the syntect-plugin.

This implementation is not quite as fancy as Sublime's; we only use the
two most basic regex patterns, for 'increase next' and 'decrease
current' indentation level.

It is, however, actual honest-to-goodness syntax-aware auto-indent,
which is a step in the right direction.

Closes #424
Progress on #937